fix(parser): report TS1520 for a v-mode class-set operator missing its left operand
The v-flag class-set operators `--` (subtraction) and `&&` (intersection)
are binary, so tsc reports TS1520 "Expected a class set operand" for a
missing operand on either side. tsz reported the missing right operand
(`/[a&&]/v`) but silently dropped the missing left operand — a class that
opens on the operator or has it right after `^` (`/[--a]/v`, `/[&&a]/v`,
`/[^--a]/v`, `/[--[a]]/v`), and under-counted `/[&&]/v`/`/[--]/v` (one
report instead of two).

When a v-mode class reaches a `--`/`&&` operator with no operand scanned
yet, emit TS1520 at the operator's first character. `committed.is_none()`
fires it once per class so a leading operator run keeps tsc's single
leading report, while the existing missing-right report is untouched.

To keep `state_expressions_literals_regex.rs` under its arch-size ceiling,
the self-contained `regex_range_order_errors` pass and its private
`parse_hex_u32` helper move into a new child module `range_order.rs` (pure
file-organization move); the duplicated `decode_surrogate_pair` collapses
into one shared module-scope const fn. Ratchet lowered 2438 -> 2191.

Fixes #17102.
