+++ crates/tsz-parser/src/parser/mod.rs
@@ -331,6 +331,10 @@ mod regex_class_set_operator_mixing_tests;
 mod regex_class_set_bare_hyphen_tests;
 
 #[cfg(test)]
+#[path = "../../tests/regex_class_set_operand_missing_tests.rs"]
+mod regex_class_set_operand_missing_tests;
+
+#[cfg(test)]
 #[path = "../../tests/modifier_ordering_tests.rs"]
 mod modifier_ordering_tests;
